refactor: move server&error to src dir
